+++ src/org/hsqldb/SessionInterface.java
@@ -1,4 +1,4 @@
-/* Copyright (c) 2001-2025, The HSQL Development Group
+/* Copyright (c) 2001-2026, The HSQL Development Group
  * All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
@@ -131,7 +131,7 @@ public interface SessionInterface {
 
     long getId();
 
-    int getRandomId();
+    long getRandomId();
 
     void resetSession();
